move ambiguous version.py files
