+++ opentelemetry-exporter-cloud-trace/tests/test_cloud_trace_exporter.py
@@ -38,9 +38,7 @@ from opentelemetry.exporter.cloud_trace import (
     _strip_characters,
     _truncate_str,
 )
-from opentelemetry.exporter.google.version import (
-    __version__ as google_ext_version,
-)
+from opentelemetry.exporter.cloud_trace.version import __version__
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import Event
 from opentelemetry.sdk.trace import _Span as Span
@@ -90,7 +88,7 @@ class TestCloudTraceSpanExporter(unittest.TestCase):
                 _strip_characters(
                     pkg_resources.get_distribution("opentelemetry-sdk").version
                 ),
-                _strip_characters(google_ext_version),
+                _strip_characters(__version__),
             )
         )
         cls.example_trace_id = "6e0c63257de34c92bf9efcd03927272e"
